requirements.txt
absl-py==2.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.51.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.63.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
loguru==0.7.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.0.3
pillow==10.3.0
protobuf==5.26.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.18.1
